doctr backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-doctr==0.6.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0 